update bspwmrc for names of application
Added a comment for Application rules and finding the XDG class names which are to be used in the rules to function properly
I used "brave-browser-stable" initially which was not functioning.  
After starting the brave browser from menu and then opened a terminal in the same desktop. Ran the command xprop | grep WM_CLASS which gives a cross in place of mouse cursor. Click anywhere inside the brave-browser window  generated the name "brave-browser", "Brave-browser". I used "Brave-browser" (without " ") in the bspwmrc application rules and voila, its done.